ui: improved empire UI sizing, JS clip queue, and mods event variant
Derive empire window bounds from the map texture and finish_pos in both
JS and C++ so layout stays consistent when assets change.

Replace hardcoded EMPIRE_SIZE in draw_background with image_get(image) dimensions.

+++ src/content/mods.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstdint>
+
 #include "core/xstring.h"
 #include "content/dir.h"
 #include "core/flat_map.h"
@@ -7,7 +9,7 @@
 #include "core/hvector.h"
 #include "core/archive.h"
 
-struct event_mods_info_updated { size_t count; };
+struct event_mods_info_updated { uint64_t count; };
 
 struct mod_entry {
     xstring name;
